+++ src/input/interactive/GenericFactory.ts
