+++ src/input/interactive/applications/index.ts
@@ -1,2 +1,2 @@
-export * from './application';
+export * from './Application';
 export * from './Modal';
